pyproject.toml
[tool.poetry]
name = "tox-automation"
version = "0.1.0"
description = ""
authors = ["Gulden Bilgutay"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
tox = "^4.11.3"
pyright = "^1.1.335"
fastapi = "^0.104.1"
ruff = "^0.1.5"
aiosqlite = "^0.19.0"
langchain = "^0.0.335"
langchain-experimental = "^0.0.40"
python-dotenv = "^1.0.0"
pytest = "^7.4.3"
httpx = "^0.25.1"
pandas = "^2.1.3"
openai = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
